+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Updated
+- Updated flow with HelperLibrary for Android (search for helper elements also by XPath). [Trello 2633](https://trello.com/c/dQqN02TN/)
+
 ## [3.210.6] - 2021-12-10
 ### Fixed
 - Improve usage of `POSITION_PROVIDER_EXTRA_SCROLL` feature in calculating scroll position. [Trello 2751](https://trello.com/c/v03eizvV)